+++ CHANGELOG.md
@@ -3,6 +3,8 @@
  - chore: Update `google_fonts` to latest
  - feat: Allow NesBottomSheet to have absolute height
  - feat: Add `NesIcons.cisel`.
+ - feat: Add `NesIcons.eyedropper`.
+ - feat: Add the `NesPalette` widget.
 
 # 0.30.0
  - feat: Add `NesShakeEffect` widget with declarative `shaking` bool and optional `NesShakeEffectController`.